Move makeString functions into own module
