+++ index.js
@@ -1,8 +1,10 @@
 /* @flow */
 import parse from './src/CssParser';
 import * as Transforms from './src/Transforms';
+import * as Strings from './src/Strings';
 
 export default {
   parse,
+  ...Strings,
   ...Transforms
 };
